Create mov_avg_filters.py
I added a few filters, in particular:
mov_avg is the same of the package provided, but it's possible to specify how many samples before and how many after.
Once called:
- if nans are contained, a loop is executed, which slows the execution for long arrays
- if no nans are contained, a scipy optimized function is called

mov_avg_time works calling mov_avg but instead of computing manually the number of samples, it's possible to provide the time array and to declare the width of the window (should be in the same measurement units). Again, it's possible to decide how many time units before and after

mov_avg_time_variable_freq works for signals collected at a sampling frequency in which the time difference between two acquisitions is relevant. Consequently, for every moment, an evaluation of which samples are inside the timing window is done.